requirements.txt
python-dotenv
groq
requests
Flask_cors
Flask
langchain==0.1.16
langchain-core
langchain-groq
gunicorn
beautifulsoup4
markdown